+++ rockpool/nn/modules/sinabs/lif_exodus.py
@@ -20,7 +20,7 @@ from sinabs.exodus.leaky import LeakyIntegrator
 
 from sinabs.activation import Heaviside, SingleExponential
 
-__all__ = ["LIFExodus", "LIFMembraneExodus", "LIFSlayer"]
+__all__ = ["LIFExodus", "LIFMembraneExodus", "LIFSlayer", "ExpSynExodus"]
 
 
 class LIFExodus(LIFBaseTorch):
@@ -47,9 +47,9 @@ class LIFExodus(LIFBaseTorch):
         Warnings:
             Exodus does not currently support training time constants, biases or thresholds. Use :py:class:`LIFTorch` or py:class:`LIFJax` to train those parameters.
 
-            Exodus currently only supports a single membrane time constant.
+            Exodus currently only supports a single threshold for the layer.
 
-            Exodus does not support neuron biases.
+            Exodus does not currently support training thresholds.
 
             Exodus does not support noise injection.
 
@@ -66,6 +66,7 @@ class LIFExodus(LIFBaseTorch):
             shape (tuple): The shape of this module
             tau_syn (float): An optional array with concrete initialisation data for the synaptic time constants. If not provided, 50ms will be used by default.
             tau_mem (float): An optional array with concrete initialisation data for the membrane time constants. If not provided, 20ms will be used by default.
+            bias (float):
             threshold (float): An optional array specifying the firing threshold of each neuron. If not provided, ``1.`` will be used by default.
             learning_window (float): Cutoff value for the surrogate gradient. Default: 0.5
             dt (float): Time step in seconds. Default: 1 ms.
@@ -87,7 +88,7 @@ class LIFExodus(LIFBaseTorch):
             tau_syn=tau_syn,
             tau_mem=tau_mem,
             threshold=Constant(threshold),
-            bias=Constant(bias),
+            bias=bias,
             has_rec=False,
             noise_std=0.0,
             learning_window=learning_window,
@@ -97,6 +98,13 @@ class LIFExodus(LIFBaseTorch):
 
         # - Assign the surrogate gradient function
         self.spike_generation_fn = Heaviside(self.learning_window)
+
+        # - Check that CUDA is available
+        if not torch.cuda.is_available():
+            raise EnvironmentError("CUDA is required for exodus-backed modules.")
+
+        # - Move module to CUDA device
+        self.cuda()
 
     def forward(self, data: torch.Tensor) -> torch.Tensor:
         """
@@ -154,13 +162,18 @@ class LIFExodus(LIFBaseTorch):
             isyn.flatten().contiguous(),  # initial state
         ).reshape(-1, self.n_synapses, time_steps)
 
-        # - For bias, add simply to isyn_exodus
-        # bias = torch.tensor(self.bias).expand((n_batches, time_steps, self.n_neurons, self.n_synapses)).flatten().contiguous()
-        # isyn_exodus += self.bias
+        # Add bias to isyn_exodus, to be added onto the membrane
+        bias = self.bias.reshape((1, -1, 1, 1))
+        bias = (
+            bias.expand((n_batches, self.n_neurons, self.n_synapses, time_steps))
+            .flatten(0, 1)
+            .contiguous()
+        )
+        isyn_with_bias = isyn_exodus + bias
 
         # Membrane dynamics: Calculate v_mem
         spikes, vmem_exodus = IntegrateAndFire.apply(
-            isyn_exodus.sum(1).contiguous(),  # input
+            isyn_with_bias.sum(1).contiguous(),  # input
             alpha.contiguous(),  # alpha
             vmem.flatten().contiguous(),  # init state
             threshold,  # threshold
@@ -201,6 +214,148 @@ class LIFExodus(LIFBaseTorch):
         return self._record_dict["spikes"]
 
 
+class ExpSynExodus(LIFBaseTorch):
+    def __init__(
+        self,
+        shape: tuple,
+        tau: P_float = 0.05,
+        noise_std: P_float = 0.0,
+        dt: float = 1e-3,
+        *args,
+        **kwargs,
+    ):
+        """
+        Instantiate an exponential synapse module using the Exodus backend
+
+        Uses the Exodus accelerated CUDA module to implement an exponential synapse. A CUDA device is required to instantiate this module.
+
+        The output of evolving this module is the synaptic currents.
+
+        Warning:
+            Exodus does not support noise injection.
+
+        Examples:
+            Instantitate an exponential synapse module with 2 synapses.
+
+            >>> mod = LIFExodus(2)
+
+            Specify the synaptic time constants, as well as time-step ``dt``.
+
+            >>> mod = LIFExodus(2, tau_syn = 10e-3, dt = 10e-3)
+
+            Specify multiple synaptic time constants.
+
+            >>> mod = LIFExodus(2, tau_syn = [10e-3, 20e-3])
+
+
+        Args:
+            shape (tuple): The shape of this module
+            tau_syn (float): An optional array with concrete initialisation data for the synaptic time constants. If not provided, 50ms will be used by default.
+            dt (float): Time step in seconds. Default: 1 ms.
+        """
+
+        # - Remove unused parameters
+        unused_arguments = ["threshold", "has_rec", "noise_std", "bias", "tau_mem"]
+        test_args = [arg in kwargs for arg in unused_arguments]
+        if any(test_args):
+            error_args = [arg for (arg, t) in zip(unused_arguments, test_args) if t]
+            raise TypeError(
+                f"The argument(s) {error_args} is/are not used in ExpSynExodus."
+            )
+
+        if noise_std != 0.0:
+            raise ValueError("`ExpSynExodus` does not support injected noise.")
+
+        # - Initialise superclass
+        super().__init__(
+            shape=shape,
+            tau_syn=tau,
+            has_rec=False,
+            noise_std=0.0,
+            dt=dt,
+            *args,
+            **kwargs,
+        )
+
+        # - Remove LIFBaseTorch attributes that do not apply
+        delattr(self, "tau_mem")
+        delattr(self, "vmem")
+        delattr(self, "threshold")
+        delattr(self, "bias")
+        delattr(self, "learning_window")
+        delattr(self, "spikes")
+        delattr(self, "spike_generation_fn")
+        delattr(self, "max_spikes_per_dt")
+
+        # - Check that CUDA is available
+        if not torch.cuda.is_available():
+            raise EnvironmentError("CUDA is required for exodus-backed modules.")
+
+        # - Move module to CUDA device
+        self.cuda()
+
+    def forward(self, data: torch.Tensor) -> torch.Tensor:
+        """
+        Forward method for processing data through this layer
+        Adds inputs to the synaptic states
+
+        ----------
+        data: Tensor
+            Data takes the shape of (batch, time_steps, n_synapses)
+
+        Returns
+        -------
+        out: Tensor
+            Out of spikes with the shape (batch, time_steps, n_synapses)
+
+        """
+        # - Ensure input data is on GPU
+        if not data.is_cuda:
+            warnings.warn("Input data was not on a CUDA device. Moving it there now.")
+        data = data.to("cuda")
+
+        # - Replicate data and states out by batches
+        data, (isyn,) = self._auto_batch(data, (self.isyn,))
+
+        # - Get input data size
+        (n_batches, time_steps, n_connections) = data.shape
+
+        # - Broadcast parameters to full size for this module
+        beta = (
+            self.calc_beta()
+            .expand((n_batches, self.n_neurons, self.n_synapses))
+            .flatten()
+        )
+
+        # Bring data into format expected by exodus: (batches*neurons*synapses, timesteps)
+        data = data.movedim(1, -1).flatten(0, -2)
+
+        # Decay data by one timestep to match xylo behavior
+        data = beta.unsqueeze(-1) * data
+
+        # Synaptic dynamics: Calculate I_syn and bring to shape
+        # (batches*neurons, synapses, timesteps)
+        isyn_exodus = LeakyIntegrator.apply(
+            data.contiguous(),  # Input
+            beta.contiguous(),  # beta
+            isyn.flatten().contiguous(),  # initial state
+        ).reshape(-1, self.n_synapses, time_steps)
+
+        # Bring states to rockpool dimensions
+        isyn_exodus = (
+            isyn_exodus.reshape(n_batches, self.n_neurons, self.n_synapses, time_steps)
+            .movedim(-1, 1)
+            .to(data.device)
+        )
+
+        # Save synaptic currents and return
+        self._record_dict["isyn"] = isyn_exodus.reshape(
+            n_batches, time_steps, self.size_out
+        )
+        self.isyn = isyn_exodus[0, -1].detach()
+        return self._record_dict["isyn"]
+
+
 class LIFMembraneExodus(LIFBaseTorch):
     def __init__(
         self,
@@ -219,8 +374,6 @@ class LIFMembraneExodus(LIFBaseTorch):
         The output of evolving this module is the neuron membrane potentials; synaptic currents are available using the ``record = True`` argument to :py:meth:`~.LIFExodus.evolve`.
 
         Warnings:
-            Exodus does not currently support training biases or thresholds.
-
             Exodus does not support noise injection.
 
         Examples:
@@ -262,7 +415,6 @@ class LIFMembraneExodus(LIFBaseTorch):
 
         # - Remove LIFBaseTorch attributes that do not apply
         delattr(self, "threshold")
-        delattr(self, "bias")
         delattr(self, "learning_window")
         delattr(self, "spikes")
         delattr(self, "spike_generation_fn")
@@ -325,6 +477,16 @@ class LIFMembraneExodus(LIFBaseTorch):
             isyn.flatten().contiguous(),  # initial state
         ).reshape(-1, self.n_synapses, time_steps)
 
+        # Add bias to isyn_exodus, to be added onto the membrane
+        bias = self.bias.reshape((1, -1, 1, 1))
+        bias = (
+            bias.expand((n_batches, self.n_neurons, self.n_synapses, time_steps))
+            .flatten(0, 1)
+            .contiguous()
+        )
+        isyn_exodus = isyn_exodus + bias
+
+        # Inteagrate onto a membrane
         vmem_exodus = LeakyIntegrator.apply(
             isyn_exodus.sum(1).contiguous(),  # input
             alpha.contiguous(),  # alpha